Add support for Consul-based deploys (#35)
We want to speed up deploys and use the Consul service registry as a
means to check health, since it updates faster than ASG+ALBs.

This also fixes a bug where green text would continue to be emitted
after the plan/apply phase was completed.

To get this to work, we also add:
- a `deploy.orchestrator` option to switch from the default (ASG-based)
to Consul
- a `deploy.healthCheckService` option to specify the name of the Consul
service to check
- logic to check the status of a Consul-based service
- a pre-container-shutdown hook that allows us to gracefully deregister
the Consul node before shutdown
- logic to create + delete ASGs outside of Terraform using the AWS API
directly (this is much faster than waiting on Terraform which waits
until the operation has fully completed, whereas we just want to kick it
off and move on)
- better handling of the `RELEASE_ID` to pass it along via environment
variable if it has been set, ensuring consistently across spawned
subprocesses
- migration-specific code to support projects migrating from ASG-based
to Consul-based deploys